refactor(models): 提取符号引脚和形状序列化器到独立文件
- 创建新的 SymbolPinSerializer 和 SymbolShapeSerializer 类来处理
  符号引脚和形状类型的 JSON 序列化
- 将 SymbolDataSerializer 中的引脚和形状序列化逻辑迁移到对应
  的专用序列化器中
- 更新 CMakeLists.txt 添加新源文件
- 重构 SymbolDataSerializer 使用委托模式调用专门的序列化器
