+++ src/models/CMakeLists.txt
@@ -4,6 +4,8 @@ add_library(EasyKiConverterModels STATIC
     ComponentData.cpp
     SymbolData.cpp
     SymbolDataSerializer.cpp
+    SymbolPinSerializer.cpp
+    SymbolShapeSerializer.cpp
     FootprintData.cpp
     FootprintDataSerializer.cpp
     Model3DData.cpp
