+++ src/services/export/CMakeLists.txt
@@ -15,6 +15,7 @@ set(EXPORT_HEADERS
     ${CMAKE_CURRENT_SOURCE_DIR}/PreviewImagesExportWorker.h
     ${CMAKE_CURRENT_SOURCE_DIR}/DatasheetExportWorker.h
     ${CMAKE_CURRENT_SOURCE_DIR}/ParallelExportService.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/ExportReportGenerator.h
     ${CMAKE_CURRENT_SOURCE_DIR}/ComponentDataCache.h
     ${CMAKE_CURRENT_SOURCE_DIR}/TempFileManager.h
     PARENT_SCOPE
@@ -33,6 +34,7 @@ set(EXPORT_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/PreviewImagesExportWorker.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/DatasheetExportWorker.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/ParallelExportService.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/ExportReportGenerator.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/ComponentDataCache.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/ExportWorkerHelpers.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/TempFileManager.cpp
